Fix type

+++ src/Tether_02.jl
@@ -2,7 +2,7 @@
 using ModelingToolkit, OrdinaryDiffEq, PyPlot, LinearAlgebra
 
 G_EARTH  = Float64[0.0, 0.0, -9.81]    # gravitational acceleration [m/s²]
-L0 = -10.0                             # initial spring length      [m]
+L0::Float64 = -10.0                    # initial spring length      [m]
 
 # model, Z component upwards
 @parameters mass=1.0 c_spring=50.0 damping=0.5 l0=L0
